move spacing into particle offset remember function
