+++ .idea/deploymentTargetDropDown.xml
@@ -6,12 +6,12 @@
         <type value="RUNNING_DEVICE_TARGET" />
         <deviceKey>
           <Key>
-            <type value="SERIAL_NUMBER" />
-            <value value="R58R72RR6XK" />
+            <type value="VIRTUAL_DEVICE_PATH" />
+            <value value="'/Users/godwinaddy/.android/avd/Pixel_6_Pro_API_33.avd'" />
           </Key>
         </deviceKey>
       </Target>
     </runningDeviceTargetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2023-05-30T10:03:30.025487Z" />
+    <timeTargetWasSelectedWithDropDown value="2023-05-30T10:11:15.108844Z" />
   </component>
 </project>
