+++ loaders/src/main/java/io/eyram/loaders/CircularLoadingAnimation.kt
@@ -17,6 +17,7 @@ import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import io.eyram.loaders.util.toComposeAngle
 import kotlinx.coroutines.launch
 
 @Composable
